+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The changelog is available [on GitHub][2].
 
 * [#459](https://github.com/kowainik/summoner/issues/459):
   Bump up `tomland` version to `1.3.0.0`.
+* [#460](https://github.com/kowainik/summoner/issues/460):
+  Bump up `relude` version to `0.7.0.0`.
+  Remove `Summoner.Template.Mempty` as `memptyIfFalse` is imported from
+  `Relude`.
 * [#455](https://github.com/kowainik/summoner/issues/455):
   Allow `validation-selective` version `0.1.0.0`.
 * [#452](https://github.com/kowainik/summoner/issues/452):